AIs align correctly and kick the ball!
- Added a set of tolerances and timers to make the kicking kind of
realistic. Without proper physics it still looks a bit awkward but this
is the basis of the AI behaviour.
- Fixed a small bug/crash regarding angle calculations but the Issue
regarding turning the longer way is still a problem
- Kicking is an instantaneous state (Then 1 second recharge), that is
sent to the Agent class, from which the Main process (Simulator) picks
up the event and then calls the Ball methods given the parameters of the
Agent that kicked the ball. There may be a more efficient way or a more
structurally consistent one, but this one works and makes sense